webkit: build JSC for i586 (JIT + C Loop) with 32-bit bindings fixes
Make WebKit/JavaScriptCore build and link for 32-bit x86:

- webkit.ts: JIT+CLOOP for i586, x86 backend, deps wiring.
- vendor-patches/WebKit: DOMWrapperWorld / JSDOMConstructorBase /
  JSDOMWrapper header fixes for 32-bit.
- jsc/bindings: fix pointer-width and x86-specific issues across
  ObjectBindings, NodeVM, v8 shim, crypto jobs, GCOutputConstraint,
  ScriptExecutionContext, CPU features, xxhash3, message port, wasm
  streaming compiler, etc.
- bundle-functions.ts codegen for 32-bit; BakeGlobalObject fixes.

+++ src/jsc/bindings/node/crypto/CryptoDhJob.cpp
@@ -111,12 +111,12 @@ std::optional<DhJobCtx> DhJobCtx::fromJS(JSGlobalObject* globalObject, ThrowScop
     const KeyObject& publicKey = publicKeyObject->handle();
 
     if (privateKey.type() != CryptoKeyType::Private) {
-        ERR::CRYPTO_INVALID_KEY_OBJECT_TYPE(scope, globalObject, privateKey.type(), "private"_s);
+        ERR::CRYPTO_INVALID_KEY_OBJECT_TYPE(scope, globalObject, static_cast<uint8_t>(privateKey.type()), "private"_s);
         return std::nullopt;
     }
 
     if (publicKey.type() != CryptoKeyType::Public && publicKey.type() != CryptoKeyType::Private) {
-        ERR::CRYPTO_INVALID_KEY_OBJECT_TYPE(scope, globalObject, publicKey.type(), "private or public"_s);
+        ERR::CRYPTO_INVALID_KEY_OBJECT_TYPE(scope, globalObject, static_cast<uint8_t>(publicKey.type()), "private or public"_s);
         return std::nullopt;
     }
